Cleaned up rendering a bit.
